Moved axis default options into AxisDefaults.
