+++ ts/Extensions/DragPanes.ts
@@ -34,6 +34,7 @@ const {
     hasTouch
 } = H;
 import Axis from '../Core/Axis/Axis.js';
+import AxisDefaults from '../Core/Axis/AxisDefaults.js';
 import palette from '../Core/Color/Palette.js';
 import Pointer from '../Core/Pointer.js';
 import U from '../Core/Utilities.js';
@@ -771,7 +772,7 @@ wrap(Pointer.prototype, 'drag', function (
     }
 });
 
-merge(true, Axis.defaultYAxisOptions, AxisResizer.resizerOptions);
+merge(true, AxisDefaults.defaultYAxisOptions, AxisResizer.resizerOptions);
 
 H.AxisResizer = AxisResizer as any;
 export default H.AxisResizer;
